+++ src/getFCP.ts
@@ -15,25 +15,30 @@
  */
 
 import {bindReporter} from './lib/bindReporter.js';
+import {finalMetrics} from './lib/finalMetrics.js';
 import {getFirstHidden} from './lib/getFirstHidden.js';
 import {initMetric} from './lib/initMetric.js';
 import {observe} from './lib/observe.js';
+import {onBFCacheRestore} from './lib/onBFCacheRestore.js';
 import {ReportHandler} from './types.js';
 
 
-export const getFCP = (onReport: ReportHandler) => {
-  const metric = initMetric('FCP');
+export const getFCP = (onReport: ReportHandler, reportAllChanges?: boolean) => {
   const firstHidden = getFirstHidden();
-
+  let metric = initMetric('FCP');
   let report: ReturnType<typeof bindReporter>;
 
   const entryHandler = (entry: PerformanceEntry) => {
     if (entry.name === 'first-contentful-paint') {
+      if (po) {
+        po.disconnect();
+      }
+
       // Only report if the page wasn't hidden prior to the first paint.
       if (entry.startTime < firstHidden.timeStamp) {
         metric.value = entry.startTime;
-        metric.isFinal = true;
         metric.entries.push(entry);
+        finalMetrics.add(metric);
         report();
       }
     }
@@ -41,6 +46,18 @@ export const getFCP = (onReport: ReportHandler) => {
 
   const po = observe('paint', entryHandler);
   if (po) {
-    report = bindReporter(onReport, metric, po);
+    report = bindReporter(onReport, metric, reportAllChanges);
+
+    onBFCacheRestore((event) => {
+      metric = initMetric('FCP');
+      report = bindReporter(onReport, metric, reportAllChanges);
+      requestAnimationFrame(() => {
+        requestAnimationFrame(() => {
+          metric.value = performance.now() - event.timeStamp;
+          finalMetrics.add(metric);
+          report();
+        });
+      });
+    });
   }
 };