+++ src/getFID.ts
@@ -15,73 +15,60 @@
  */
 
 import {bindReporter} from './lib/bindReporter.js';
+import {finalMetrics} from './lib/finalMetrics.js';
 import {getFirstHidden} from './lib/getFirstHidden.js';
 import {initMetric} from './lib/initMetric.js';
 import {observe, PerformanceEntryHandler} from './lib/observe.js';
+import {onBFCacheRestore} from './lib/onBFCacheRestore.js';
 import {onHidden} from './lib/onHidden.js';
-import {ReportHandler} from './types.js';
+import {firstInputPolyfill, resetFirstInputPolyfill} from './lib/polyfills/firstInputPolyfill.js';
+import {FirstInputPolyfillCallback, PerformanceEventTiming, ReportHandler} from './types.js';
 
 
-interface FIDPolyfillCallback {
-  (value: number, event: Event): void;
-}
-
-interface FIDPolyfill {
-  onFirstInputDelay: (onReport: FIDPolyfillCallback) => void;
-}
-
-declare global {
-  interface Window {
-    perfMetrics: FIDPolyfill;
-  }
-}
-
-// https://wicg.github.io/event-timing/#sec-performance-event-timing
-interface PerformanceEventTiming extends PerformanceEntry {
-  processingStart: DOMHighResTimeStamp;
-  cancelable?: boolean;
-  target?: Element;
-}
-
-export const getFID = (onReport: ReportHandler) => {
-  const metric = initMetric('FID');
+export const getFID = (onReport: ReportHandler, reportAllChanges?: boolean) => {
   const firstHidden = getFirstHidden();
+  let metric = initMetric('FID');
+  let report: ReturnType<typeof bindReporter>;
 
   const entryHandler = (entry: PerformanceEventTiming) => {
     // Only report if the page wasn't hidden prior to the first input.
     if (entry.startTime < firstHidden.timeStamp) {
       metric.value = entry.processingStart - entry.startTime;
       metric.entries.push(entry);
-      metric.isFinal = true;
+      finalMetrics.add(metric);
       report();
     }
   };
 
   const po = observe('first-input', entryHandler as PerformanceEntryHandler);
-  const report = bindReporter(onReport, metric, po);
+  report = bindReporter(onReport, metric, reportAllChanges);
 
   if (po) {
     onHidden(() => {
       po.takeRecords().map(entryHandler as PerformanceEntryHandler);
       po.disconnect();
     }, true);
+  }
+
+  if (self.__WEB_VITALS_POLYFILL__) {
+    // Prefer the native implementation if available,
+    if (!po) {
+      window.webVitals.firstInputPolyfill(entryHandler as FirstInputPolyfillCallback)
+    }
+    onBFCacheRestore(() => {
+      metric = initMetric('FID');
+      report = bindReporter(onReport, metric, reportAllChanges);
+      window.webVitals.resetFirstInputPolyfill();
+      window.webVitals.firstInputPolyfill(entryHandler as FirstInputPolyfillCallback);
+    });
   } else {
-    if (window.perfMetrics && window.perfMetrics.onFirstInputDelay) {
-      window.perfMetrics.onFirstInputDelay((value: number, event: Event) => {
-        // Only report if the page wasn't hidden prior to the first input.
-        if (event.timeStamp < firstHidden.timeStamp) {
-          metric.value = value;
-          metric.isFinal = true;
-          metric.entries = [{
-            entryType: 'first-input',
-            name: event.type,
-            target: event.target,
-            cancelable: event.cancelable,
-            startTime: event.timeStamp,
-            processingStart: event.timeStamp + value,
-          } as PerformanceEventTiming];
-          report();
-        }
+    // Only monitor bfcache restores if the browser supports FID natively.
+    if (po) {
+      onBFCacheRestore(() => {
+        metric = initMetric('FID');
+        report = bindReporter(onReport, metric, reportAllChanges);
+        resetFirstInputPolyfill();
+        firstInputPolyfill(entryHandler as FirstInputPolyfillCallback);
       });
     }
   }