+++ src/lib/bindReporter.ts
@@ -14,23 +14,20 @@
  * limitations under the License.
  */
 
+import {finalMetrics} from './finalMetrics.js';
 import {Metric, ReportHandler} from '../types.js';
 
 
 export const bindReporter = (
   callback: ReportHandler,
   metric: Metric,
-  po: PerformanceObserver | undefined,
-  observeAllUpdates?: boolean,
+  reportAllChanges?: boolean,
 ) => {
   let prevValue: number;
   return () => {
-    if (po && metric.isFinal) {
-      po.disconnect();
-    }
     if (metric.value >= 0) {
-      if (observeAllUpdates ||
-          metric.isFinal ||
+      if (reportAllChanges ||
+          finalMetrics.has(metric) ||
           document.visibilityState === 'hidden') {
         metric.delta = metric.value - (prevValue || 0);
 
@@ -38,9 +35,9 @@ export const bindReporter = (
         // final, or if no previous value exists (which can happen in the case
         // of the document becoming hidden when the metric value is 0).
         // See: https://github.com/GoogleChrome/web-vitals/issues/14
-        if (metric.delta || metric.isFinal || prevValue === undefined) {
-          callback(metric);
+        if (metric.delta || prevValue === undefined) {
           prevValue = metric.value;
+          callback(metric);
         }
       }
     }