+++ test/e2e/getFID-test.js
@@ -17,6 +17,7 @@
 const assert = require('assert');
 const {beaconCountIs, clearBeacons, getBeacons} = require('../utils/beacons.js');
 const {browserSupportsEntry} = require('../utils/browserSupportsEntry.js');
+const {stubForwardBack} = require('../utils/stubForwardBack.js');
 const {stubVisibilityChange} = require('../utils/stubVisibilityChange.js');
 
 
@@ -43,11 +44,10 @@ describe('getFID()', async function() {
 
     const [fid] = await getBeacons();
     assert(fid.value >= 0);
-    assert(fid.id.match(/\d+-\d+/));
+    assert(fid.id.match(/^v1-\d+-\d+$/));
     assert.strictEqual(fid.name, 'FID');
     assert.strictEqual(fid.value, fid.delta);
     assert.strictEqual(fid.entries[0].name, 'mousedown');
-    assert.strictEqual(fid.isFinal, true);
   });
 
   it('does not report if the browser does not support FID and the polyfill is not used', async function() {
@@ -62,8 +62,20 @@ describe('getFID()', async function() {
     // Wait a bit to ensure no beacons were sent.
     await browser.pause(1000);
 
-    const beacons = await getBeacons();
-    assert.strictEqual(beacons.length, 0);
+    const loadBeacons = await getBeacons();
+    assert.strictEqual(loadBeacons.length, 0);
+
+    await stubForwardBack();
+
+    // Assert no entries after bfcache restores either (if the browser does
+    // not support native FID and the polyfill is not used).
+    await h1.click();
+
+    // Wait a bit to ensure no beacons were sent.
+    await browser.pause(1000);
+
+    const bfcacheRestoreBeacons = await getBeacons();
+    assert.strictEqual(bfcacheRestoreBeacons.length, 0);
   });
 
   it('falls back to the polyfill in non-supporting browsers', async function() {
@@ -82,10 +94,9 @@ describe('getFID()', async function() {
     const [fid] = await getBeacons();
 
     assert(fid.value >= 0);
-    assert(fid.id.match(/\d+-\d+/));
+    assert(fid.id.match(/^v1-\d+-\d+$/));
     assert.strictEqual(fid.name, 'FID');
     assert.strictEqual(fid.value, fid.delta);
-    assert.strictEqual(fid.isFinal, true);
     assert.strictEqual(fid.entries[0].name, 'mousedown');
     if (browserSupportsFID) {
       assert('duration' in fid.entries[0]);
@@ -133,6 +144,41 @@ describe('getFID()', async function() {
 
     const beacons = await getBeacons();
     assert.strictEqual(beacons.length, 0);
+  });
+
+  it('reports the first input delay after bfcache restores', async function() {
+    if (!browserSupportsFID) this.skip();
+
+    await browser.url('/test/fid');
+
+    // Click on the <h1>.
+    const h1 = await $('h1');
+    await h1.click();
+
+    await beaconCountIs(1);
+
+    const [fid1] = await getBeacons();
+    assert(fid1.value >= 0);
+    assert(fid1.id.match(/^v1-\d+-\d+$/));
+    assert.strictEqual(fid1.name, 'FID');
+    assert.strictEqual(fid1.value, fid1.delta);
+    assert.strictEqual(fid1.entries[0].name, 'mousedown');
+
+    await clearBeacons();
+    await stubForwardBack();
+
+    // Click on the <h1>.
+    await h1.click();
+
+    await beaconCountIs(1);
+
+    const [fid2] = await getBeacons();
+    assert(fid2.value >= 0);
+    assert(fid2.id.match(/^v1-\d+-\d+$/));
+    assert(fid1.id !== fid2.id);
+    assert.strictEqual(fid2.name, 'FID');
+    assert.strictEqual(fid2.value, fid2.delta);
+    assert.strictEqual(fid2.entries[0].name, 'mousedown');
   });
 });