+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v1.0.0 (2020-11-12)
+
+- **[BREAKING]** Add support for reporting metrics on back/forward cache restore ([#87](https://github.com/GoogleChrome/web-vitals/pull/87))
+
+- **[BREAKING]** Remove the `isFinal` flag from the Metric interface ([#86](https://github.com/GoogleChrome/web-vitals/pull/86))
+
+- Remove the scroll listener to stop LCP observing ([#85](https://github.com/GoogleChrome/web-vitals/pull/85))
+
 ### v0.2.4 (2020-07-23)
 
 - Remove the unload listener ([#68](https://github.com/GoogleChrome/web-vitals/pull/68))